replace the use of identity hash map with standard hashmap
IdentityHashMap causes a stackoverflow when attempting to serialize
Double.
It failed to recognise that Double.POSITIVE_INFINITY is already in the
map and continues to add it until it blows up.

I've not been able to reproduce this in unit tests, it only seems to
show up in prod. Still investigating. Using a standard HashMap which
does object as opposed to reference comparison doesn't have this issue.
See issue #4