use pooled buffer and check flush failures
ensure when a message is written that it is flushed.
if flushing fails, log and close the connection.
there is currently no other way to recover since the client uses the
first 5 bytes of data to determine the size, if a flush failure occurs
it will incorrectly prepend the old, failed message to the new one
leading to all future messages being corrupted.
This fail fast approach avoids this potential data corruption.