introduce generic file loading and resolution
The previous methods of resolving a file was error prone and broke in
many cases if the file we’re looking for is in a JAR.
This approach introduces the FileUtil and ResolvedFile classes which
facilitate a better mechanism for finding files.

This makes heavy use of the Files and Paths API and usage of the old
File class should eventually be migrated to use the new method.

File resolution works by checking if the file exists on disk. If it
doesn’t it checks if it is on the class path.
If a file is found, regardless of where, the new API provides an
InputStream to the file as the preferred method of reading. If the file
is a directory then it provides a directory stream.